% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc_utils.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute Area Under the Receiver Operating Characteristic Curve (AUC)}
\usage{
compute_auc(preds, labs)
}
\arguments{
\item{preds}{vector of predicted probabilities (numeric)}

\item{labs}{vector of true class labels (numeric)}
}
\value{
value of Area Under the Receiver Operating Characteristic Curve (AUC) (numeric)
}
\description{
Compute Area Under the Receiver Operating Characteristic Curve (AUC)
}
\seealso{
\code{\link[ROCR]{performance}}
}
