% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recidivism}
\alias{recidivism}
\title{Return status of inmates from the state of New York.}
\format{A data frame with 123588 rows and 6 variables:
#' \describe{
  \item{Release.Year}{Year the individual was released.}
  \item{County.of.Indictment}{County of New York in which the original indictment occurred.}
  \item{Gender}{Gender (factor)}
  \item{Age.at.Release}{Inmate age at time of release.}
  \item{Return.Status}{Description of return status (factor).}
}}
\source{
\url{https://catalog.data.gov/dataset/recidivism-beginning-2008}
}
\usage{
recidivism
}
\description{
Represents return status within three years of release among inmates
released to the community during a particular calendar year. Each data
record represents a release to the community as a result of completing
maximum sentence, reaching conditional release date, or being approved
for release by the Board of Parole. The dataset includes data about
release year, county of indictment, gender, age at release, and
return status.
}
\keyword{datasets}
