% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{slice_plot}
\alias{slice_plot}
\title{Create a "slice plot" of two roc curves with area between them (the ABROCA region) shaded.}
\usage{
slice_plot(majority_roc, minority_roc, majority_group_name = NULL,
  minority_group_name = NULL, fout = NULL)
}
\arguments{
\item{majority_roc}{list with attributes "x" and "y" defining points of roc curve}

\item{minority_roc}{list with attributes "x" and "y" defining points of roc curve}

\item{minority_group_name}{optional label for minority group (character)}

\item{fout}{path to output file}

\item{majoruty_group_name}{optional label for majority group (character)}
}
\value{
No return value; file is saved to disk.
}
\description{
Create a "slice plot" of two roc curves with area between them (the ABROCA region) shaded.
}
