% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc_utils.R
\name{compute_roc}
\alias{compute_roc}
\title{compute the Receiver Operating Characteristic (ROC) curve for a set of predicted probabilities and the true class labels.}
\usage{
compute_roc(preds, labs)
}
\arguments{
\item{preds}{vector of predicted probability of being in the positive class P(X == 1) (numeric)}

\item{labs}{vector of true labels (numeric)}
}
\value{
ROCR::performance object
}
\description{
compute the Receiver Operating Characteristic (ROC) curve for a set of predicted probabilities and the true class labels.
}
\seealso{
\code{\link{ROCR::performance}}
}
