% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc_utils.R
\name{interpolate_roc_fun}
\alias{interpolate_roc_fun}
\title{Use interpolation to make approximate the Receiver Operating Characteristic (ROC) curve along n_grid equally-spaced values.}
\usage{
interpolate_roc_fun(perf_in, n_grid = 10000)
}
\arguments{
\item{perf_in}{ROCR::performance object computed via \code{\link{compute_roc}}}

\item{n_grid}{number of approximation points to use (default value of 10000 more than adequate for most applications) (numeric)}
}
\value{
returns a list with components x and y, containing n coordinates which
  interpolate the given data points according to the method (and rule) desired.
}
\description{
Use interpolation to make approximate the Receiver Operating Characteristic (ROC) curve along n_grid equally-spaced values.
}
\seealso{
\code{\link[stats]{approx}}
}
