% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{roc_plot}
\alias{roc_plot}
\title{Plot the Receiver Operating Characteristic curve for preds and labs.}
\usage{
roc_plot(preds, labs, plot_type = NULL, show_diag = FALSE,
  fout = NULL)
}
\arguments{
\item{preds}{vector of predictions}

\item{labs}{vector of true class labels}

\item{plot_type}{if set to "majority" or "minority", coloring is set to match
plots generated by slice_plot (blue for majority class, red for minority);
if NULL then neutral coloring is used}

\item{show_diag}{boolean indicator for whether a diagonal line indicating the
random chance ROC curve should be plotted}

\item{fout}{path to an output file (png)}
}
\value{
No return value; file is saved to disk.
}
\description{
Plot the Receiver Operating Characteristic curve for preds and labs.
}
\examples{
 
# First, we load data, train a model, and generate predictions to evaluate.
data("recidivism")
recidivism$returned = as.factor(recidivism$Return.Status != "Not Returned")
in_train = caret::createDataPartition(recidivism$returned, 
    p = 0.75, list = FALSE)
traindata = recidivism[in_train,c("Release.Year", "County.of.Indictment", 
    "Gender", "Age.at.Release", "returned")]
testdata = recidivism[-in_train,c("Release.Year", "County.of.Indictment", 
    "Gender", "Age.at.Release", "returned")]
lr = glm(returned ~ ., data=traindata, family="binomial")
testdata$pred = predict(lr, testdata, type = "response")

# Now, we apply roc_plot() to the labels and predictions 
# (note that this writes a file to fout):
roc_plot(testdata$pred, testdata$returned, plot_type = "majority", 
    show_diag=TRUE, fout="roc.png")
}
